refactor(cache): per-store/per-Vary SWR guard, Vary-on-freshen, split revalidation module
Addresses Copilot review on PR #54:
- backgroundRefreshes is now a per-store WeakMap keyed additionally by the
  selected Vary variant, so refreshes for different stores or variants of the
  same URL no longer suppress each other (findings 1 & 2).
- Freshening a 304 recomputes the stored `vary` selector map from the merged
  Vary header instead of keeping the stale map, so stored metadata stays
  consistent with served headers and variant matching survives an origin
  changing Vary (finding 3). Extracted a shared `parseVary` helper used by both
  the store and freshen paths.
- Moved RevalidationHandler + backgroundRefresh into a new
  interceptor/cache-revalidation.js for readability; cache.js exports the
  shared helpers it needs.

Tests: added coalescing, per-store isolation, per-Vary, Vary-on-freshen, and
failed-background-refresh cases to test/cache-revalidation.js (the finding-1/2/3
cases fail against the pre-fix code).
