+++ lib/interceptor/cache.js
@@ -10,11 +10,11 @@ import {
 } from '../utils.js'
 import { isHopByHop } from './proxy.js'
 import { SqliteCacheStore } from '../sqlite-cache-store.js'
-import { traceWrite, traceSafe, traceErr, traceUrl } from '../trace.js'
+import { RevalidationHandler, backgroundRefresh } from './cache-revalidation.js'
 
 let DEFAULT_STORE = null
 const DEFAULT_MAX_ENTRY_SIZE = 128 * 1024
-const DEFAULT_MAX_ENTRY_TTL = 30 * 24 * 3600 // seconds
+export const DEFAULT_MAX_ENTRY_TTL = 30 * 24 * 3600 // seconds
 // Bounded retention for entries kept solely for conditional revalidation
 // (zero/expired freshness but a usable validator): long enough that the
 // "always validate" origin pattern pays a 304 instead of a full 200, short
@@ -24,10 +24,6 @@ const REVALIDATION_RETENTION = 24 * 3600 // seconds
 // 1-year default, capped by maxEntryTTL below.
 const IMMUTABLE_LIFETIME = 31556952 // seconds
 const NOOP = () => {}
-
-// In-flight background stale-while-revalidate refreshes keyed by
-// method + url, so a hot stale key spawns one refresh, not a herd.
-const backgroundRefreshes = new Set()
 
 /**
  * Explicit (or opt-in heuristic) freshness lifetime in seconds, or null when
@@ -40,7 +36,7 @@ const backgroundRefreshes = new Set()
  *
  * @returns {{ lifetime: number, explicit: boolean } | null}
  */
-function determineLifetime(
+export function determineLifetime(
   statusCode,
   headers,
   cacheControlDirectives,
@@ -100,7 +96,7 @@ function determineLifetime(
  * origin Date). A response relayed through intermediaries that don't add Age
  * would otherwise get an over-extended TTL and be served stale.
  */
-function determineAge(headers, now) {
+export function determineAge(headers, now) {
   const rawAge = headers.age
   // A duplicated Age header arrives as an array; take the first value.
   const rawAgeValue = Array.isArray(rawAge) ? rawAge[0] : rawAge
@@ -131,7 +127,7 @@ function determineAge(headers, now) {
  * validator-bearing entries; everything is capped by maxEntryTTL, measured
  * from (backdated) cachedAt.
  */
-function computeEntryTimes(
+export function computeEntryTimes(
   lifetime,
   explicit,
   age,
@@ -187,6 +183,43 @@ function forbidsServingStale(entry) {
 }
 
 /**
+ * Builds the Vary selector map (RFC 9111 §4.1): for each field named in the
+ * response Vary header, records the request's value (a null sentinel when the
+ * header was absent — absent-vs-present is a mismatch, so an empty map must
+ * NOT act as a wildcard). Selector names are lowercased and stored on a null
+ * prototype: names come from the (server-controlled) Vary header, so a
+ * `__proto__` entry on a plain `{}` would hit the Object.prototype setter and
+ * be silently dropped.
+ *
+ * Returns null when the response is NOT cacheable on Vary grounds: a
+ * non-string Vary (duplicated header lines) or a Vary containing '*' (which
+ * never matches, RFC 9111 §4.1).
+ *
+ * @returns {Record<string, string | string[] | null> | null}
+ */
+export function parseVary(varyHeader, requestHeaders) {
+  const vary = Object.create(null)
+  if (varyHeader == null) {
+    return vary
+  }
+  if (typeof varyHeader !== 'string') {
+    return null
+  }
+  for (const name of varyHeader.split(',').map((key) => key.trim().toLowerCase())) {
+    if (name === '*') {
+      return null
+    }
+    if (name === '') {
+      // Empty field-name (a bare/trailing comma, or `Vary:` with no value) is
+      // not a selector — skip it so it can't become a spurious '' key.
+      continue
+    }
+    vary[name] = requestHeaders[name] ?? null
+  }
+  return vary
+}
+
+/**
  * Conditional request headers for revalidating a stored entry (RFC 9111
  * §4.3.1). Per RFC 9110 §13.1.3 and undici PR #5512, If-Modified-Since echoes
  * the stored Last-Modified VERBATIM when available (nginx's default
@@ -196,7 +229,7 @@ function forbidsServingStale(entry) {
  * The stored vary values need no replay: the store only returns entries whose
  * selecting headers already match this request.
  */
-function conditionalHeaders(headers, entry) {
+export function conditionalHeaders(headers, entry) {
   const condHeaders = { ...headers }
   if (entry.etag) {
     condHeaders['if-none-match'] = entry.etag
@@ -212,7 +245,7 @@ function conditionalHeaders(headers, entry) {
   return condHeaders
 }
 
-class CacheHandler extends DecoratorHandler {
+export class CacheHandler extends DecoratorHandler {
   #key
   #value
   #store
@@ -392,27 +425,11 @@ class CacheHandler extends DecoratorHandler {
     // serve stale, ...), which the read path enforces from the persisted
     // cacheControlDirectives.
 
-    // Null prototype: selector names come from the response Vary header and
-    // request header names are caller-controlled — on a plain `{}` a
-    // `__proto__` key would hit the prototype setter and be silently dropped.
-    const vary = Object.create(null)
-    if (headers.vary) {
-      if (typeof headers.vary !== 'string') {
-        return this.#skip('vary-invalid', statusCode, headers, resume)
-      }
-
-      for (const key of headers.vary.split(',').map((key) => key.trim().toLowerCase())) {
-        if (key === '*') {
-          // RFC 9111 §4.1: a Vary field containing '*' never matches.
-          return this.#skip('vary-star', statusCode, headers, resume)
-        }
-        // Record every selecting header, using a null sentinel when it was
-        // absent from the request. RFC 9111 §4.1: absent-vs-present is a
-        // mismatch, so an empty vary object must NOT act as a wildcard that
-        // matches requests which later supply the header. headerValueEquals
-        // treats null/absent as equal only to another null/absent.
-        vary[key] = this.#key.headers[key] ?? null
-      }
+    // parseVary returns null when the response can't be cached on Vary
+    // grounds (a non-string Vary header, or one containing '*').
+    const vary = parseVary(headers.vary, this.#key.headers)
+    if (vary == null) {
+      return super.onHeaders(statusCode, headers, resume)
     }
 
     const now = Date.now()
@@ -588,325 +605,6 @@ class CacheHandler extends DecoratorHandler {
 }
 
 /**
- * Drives a conditional (revalidation) request to the origin for a stale
- * stored entry (RFC 9111 §4.3). Decision at response headers:
- * - 304: the entry is valid — freshen it (§4.3.4) and serve it.
- * - 5xx within the stale-if-error window (RFC 5861 §4, incl. undici PR
- *   #5513's pre-response connection errors): discard the error and serve the
- *   stale entry.
- * - anything else: a real replacement — stream it through a CacheHandler to
- *   the user handler, storing it on the way.
- *
- * The user handler's response callbacks are DEFERRED until that decision, but
- * onConnect is forwarded eagerly with a bridging abort — the user (e.g. a
- * signal listener in request()) must be able to tear down the in-flight
- * conditional request; nothing below the cache observes opts.signal
- * mid-flight. Delivery paths re-invoke onConnect (serveFromCache / the pass
- * CacheHandler drive their own connect), which handlers in this pipeline
- * already tolerate — the retry interceptor re-connects the same handler on
- * every attempt.
- */
-class RevalidationHandler {
-  #key
-  #entry
-  #store
-  #logger
-  #opts
-  #cacheOpts
-  #userHandler
-  #allowStaleOnError
-  #noStore
-  /** @type {null | 'pass' | 'validated' | 'stale'} */
-  #mode = null
-  /** @type {CacheHandler | import('../utils.js').DecoratorHandler | object | null} */
-  #inner = null
-  /** @type {((reason?: any) => void) | null} */
-  #abort = null
-  /** @type {Record<string, string | string[]> | null} */
-  #headers304 = null
-  #delivered = false
-  #userAborted = false
-  #userAbortReason = null
-
-  constructor(key, entry, opts, { store, logger, handler, allowStaleOnError, noStore, cacheOpts }) {
-    this.#key = key
-    this.#entry = entry
-    this.#store = store
-    this.#logger = logger
-    this.#opts = opts
-    this.#cacheOpts = cacheOpts
-    this.#userHandler = handler
-    this.#allowStaleOnError = allowStaleOnError
-    this.#noStore = noStore ?? false
-  }
-
-  onConnect(abort) {
-    this.#abort = abort
-    // Eager connect: hand the user a bridging abort so a caller abort (e.g.
-    // a signal firing in request()) tears down the in-flight conditional
-    // request instead of being silently ignored until the origin answers.
-    this.#userHandler.onConnect((reason) => {
-      if (!this.#delivered && !this.#userAborted) {
-        this.#userAborted = true
-        this.#userAbortReason = reason
-        this.#abort?.(reason)
-      }
-    })
-  }
-
-  onUpgrade(statusCode, headers, socket) {
-    // Upgrades are gated out before the cache interceptor engages; nothing to
-    // do but not crash if one slips through.
-    socket?.destroy?.()
-  }
-
-  onHeaders(statusCode, headers, resume) {
-    if (statusCode < 200) {
-      // Informational (1xx) interim responses — e.g. 103 Early Hints — are
-      // not the revalidation answer; stay undecided and keep reading.
-      return true
-    }
-
-    if (statusCode === 304) {
-      this.#mode = 'validated'
-      this.#headers304 = headers
-      return true
-    }
-
-    if (this.#allowStaleOnError && statusCode >= 500 && statusCode <= 504) {
-      this.#mode = 'stale'
-      // The decision to serve stale is already made — don't wait for the 5xx
-      // body to drain (it can be large and only adds latency/bandwidth).
-      // Deliver the stale entry now and abort the in-flight request; the
-      // resulting onError is suppressed by the #delivered guard.
-      this.#deliver(this.#entry)
-      this.#abort?.(new Error('stale-if-error: serving stale cached response'))
-      return false
-    }
-
-    this.#mode = 'pass'
-    // RFC 9111 §5.2.1.5: a request no-store forbids storing the replacement
-    // response — stream it to the user handler without the CacheHandler wrap.
-    this.#inner = this.#noStore
-      ? this.#userHandler
-      : new CacheHandler(this.#key, {
-          ...this.#cacheOpts,
-          store: this.#store,
-          logger: this.#logger,
-          handler: this.#userHandler,
-        })
-    this.#inner.onConnect((reason) => this.#abort?.(reason))
-    return this.#inner.onHeaders(statusCode, headers, resume)
-  }
-
-  onData(chunk) {
-    return this.#mode === 'pass' ? this.#inner.onData(chunk) : true
-  }
-
-  onComplete(trailers) {
-    if (this.#mode === 'pass') {
-      return this.#inner.onComplete(trailers)
-    }
-    // Freshen even when the user aborted mid-flight — the 304 validated the
-    // entry either way, and the store write benefits every other caller.
-    this.#deliver(this.#mode === 'validated' ? this.#freshen() : this.#entry)
-  }
-
-  onError(err) {
-    if (this.#mode === 'pass') {
-      return this.#inner.onError(err)
-    }
-    if (this.#userAborted) {
-      // The (probable) cause of this error is the user's own abort — a
-      // stale-if-error serve would convert their cancellation into a
-      // successful response.
-      return this.#fail(this.#userAbortReason ?? err)
-    }
-    if (this.#mode === 'validated') {
-      // The 304 already validated the entry; a broken tail on the (empty)
-      // validation response doesn't invalidate it.
-      return this.#deliver(this.#freshen())
-    }
-    if (this.#mode === 'stale' || this.#allowStaleOnError) {
-      // Pre-response connection errors (ECONNREFUSED, reset, ...) count as
-      // origin errors for stale-if-error — RFC 5861's definition explicitly
-      // covers an unreachable origin (undici PR #5513). Mid-body failures of
-      // a replacement response are 'pass' mode and propagate above.
-      return this.#deliver(this.#entry)
-    }
-    this.#fail(err)
-  }
-
-  #deliver(entry) {
-    if (this.#delivered) {
-      return
-    }
-    this.#delivered = true
-    if (this.#userAborted) {
-      this.#userHandler.onError(this.#userAbortReason ?? new Error('aborted'))
-    } else {
-      serveFromCache(entry, this.#opts, this.#userHandler)
-    }
-  }
-
-  #fail(err) {
-    if (!this.#delivered) {
-      this.#delivered = true
-      this.#userHandler.onError(err)
-    }
-  }
-
-  /**
-   * RFC 9111 §4.3.4: merge the 304's headers over the stored ones and reset
-   * the freshness clock. Returns the entry to serve; the store is only
-   * updated when the merged response is still storable.
-   */
-  #freshen() {
-    const entry = this.#entry
-    const headers304 = this.#headers304 ?? {}
-    try {
-      const now = Date.now()
-
-      // Same exclusions as at store time — hop-by-hop, fields listed in the
-      // 304's Connection header — plus Set-Cookie (must never enter a shared
-      // entry) and the body-describing fields, for which the stored body is
-      // authoritative.
-      const excludedHeaders = new Set(['age', 'set-cookie', 'content-length', 'content-range'])
-      const connection = headers304.connection
-      if (typeof connection === 'string') {
-        for (const name of connection.split(',')) {
-          excludedHeaders.add(name.trim().toLowerCase())
-        }
-      } else if (Array.isArray(connection)) {
-        for (const line of connection) {
-          for (const name of `${line}`.split(',')) {
-            excludedHeaders.add(name.trim().toLowerCase())
-          }
-        }
-      }
-
-      const merged = Object.create(null)
-      for (const name of Object.keys(entry.headers ?? {})) {
-        merged[name] = entry.headers[name]
-      }
-      for (const name of Object.keys(headers304)) {
-        const lower = name.toLowerCase()
-        if (isHopByHop(lower) || excludedHeaders.has(lower)) {
-          continue
-        }
-        const val = headers304[name]
-        merged[lower] = Array.isArray(val) ? val.slice() : val
-      }
-
-      // RFC 9110 §6.6.1: a 304 without a Date header is assigned the receipt
-      // time. Without this, the stored (old) Date drives the corrected age
-      // below and the freshened entry lands stale-on-arrival — permanently
-      // revalidating on every hit despite the origin granting freshness.
-      if (headers304.date == null) {
-        merged.date = new Date(now).toUTCString()
-      }
-
-      const cacheControlDirectives = parseCacheControl(merged['cache-control']) ?? {}
-      if (cacheControlDirectives['no-store'] || cacheControlDirectives.private === true) {
-        // The origin has withdrawn (shared-)cacheability: serve this one
-        // validated use, but don't freshen the stored entry — it stays stale
-        // and keeps revalidating until it ages out.
-        return { ...entry, headers: merged }
-      }
-
-      // Qualified no-cache=/private= field lists (possibly newly added by
-      // the 304) must not survive in the stored — or served — headers, even
-      // when the named field came from the original stored response.
-      for (const directive of ['no-cache', 'private']) {
-        if (Array.isArray(cacheControlDirectives[directive])) {
-          for (const name of cacheControlDirectives[directive]) {
-            delete merged[name]
-          }
-        }
-      }
-
-      const etag =
-        typeof merged.etag === 'string' && isEtagUsable(merged.etag)
-          ? merged.etag
-          : (entry.etag ?? '')
-      const hasValidator = etag !== '' || typeof merged['last-modified'] === 'string'
-
-      let lifetimeInfo = determineLifetime(
-        entry.statusCode,
-        merged,
-        cacheControlDirectives,
-        this.#cacheOpts,
-        now,
-      )
-      if (lifetimeInfo == null && cacheControlDirectives['no-cache'] === true) {
-        lifetimeInfo = { lifetime: 0, explicit: true }
-      }
-      if (lifetimeInfo == null) {
-        return { ...entry, headers: merged }
-      }
-
-      // Clamp the corrected age to the stored entry's own current age
-      // (now - cachedAt already includes the original initial age, RFC 9111
-      // §4.2.3): a skewed/old Date on the 304 must not push the freshened
-      // entry further into the past than the response it just validated.
-      const age = Math.min(
-        determineAge(merged, now),
-        Math.max(0, Math.floor((now - entry.cachedAt) / 1000)),
-      )
-      const times = computeEntryTimes(
-        lifetimeInfo.lifetime,
-        lifetimeInfo.explicit,
-        age,
-        cacheControlDirectives,
-        this.#cacheOpts.maxEntryTTL ?? this.#store.maxEntryTTL ?? DEFAULT_MAX_ENTRY_TTL,
-        hasValidator,
-        now,
-      )
-      if (times == null) {
-        return { ...entry, headers: merged }
-      }
-
-      const freshened = {
-        // 206 entries never take the revalidation path, so the stored body is
-        // the full representation: start 0, end = byte length (0 for HEAD).
-        body: entry.body ?? null,
-        start: 0,
-        end: entry.body ? entry.body.byteLength : 0,
-        statusCode: entry.statusCode,
-        statusMessage: entry.statusMessage ?? '',
-        headers: merged,
-        cacheControlDirectives,
-        etag,
-        vary: entry.vary,
-        cachedAt: times.cachedAt,
-        staleAt: times.staleAt,
-        deleteAt: times.deleteAt,
-      }
-
-      // RFC 9111 §5.2.1.5: a request no-store forbids storing any part of
-      // this request's response — serve the freshened value, don't write it.
-      if (!this.#noStore) {
-        try {
-          this.#store.set(this.#key, freshened)
-        } catch (err) {
-          if (err.message === 'database is locked') {
-            this.#logger?.debug({ err }, 'failed to freshen cache entry')
-          } else {
-            this.#logger?.error({ err }, 'failed to freshen cache entry')
-          }
-        }
-      }
-
-      return freshened
-    } catch (err) {
-      // Freshening must never break delivery of a validated entry.
-      this.#logger?.error({ err }, 'failed to freshen cache entry')
-      return entry
-    }
-  }
-}
-
-/**
  * RFC 9111 §4.4: a non-error response to an unsafe method invalidates the
  * stored entries for the target URI and any same-origin Location /
  * Content-Location URIs (undici PR #5514). Cross-origin targets are skipped —
@@ -1076,70 +774,12 @@ function makeKey(opts) {
   return key
 }
 
-function cacheOptsOf(opts) {
+export function cacheOptsOf(opts) {
   return {
     maxEntrySize: opts.cache.maxEntrySize,
     maxEntryTTL: opts.cache.maxEntryTTL,
     heuristic: opts.cache.heuristic,
     defaultTTL: opts.cache.defaultTTL,
-  }
-}
-
-/**
- * Fire-and-forget background refresh for stale-while-revalidate (RFC 5861
- * §3): the caller was already served the stale entry; only the store observes
- * the outcome. Re-enters the dispatch chain below the cache interceptor.
- */
-function backgroundRefresh(dispatch, opts, key, store, entry) {
-  const refreshKey = `${key.method}:${key.origin}${key.path}`
-  if (backgroundRefreshes.has(refreshKey)) {
-    return
-  }
-  backgroundRefreshes.add(refreshKey)
-
-  let finished = false
-  const done = () => {
-    if (!finished) {
-      finished = true
-      backgroundRefreshes.delete(refreshKey)
-    }
-  }
-
-  const silentHandler = {
-    onConnect: NOOP,
-    onHeaders: () => true,
-    onData: () => true,
-    onComplete: done,
-    onError: (err) => {
-      done()
-      opts.logger?.debug({ err }, 'cache: background revalidation failed')
-    },
-  }
-
-  try {
-    // Strip caller-specific request state: the body (already consumed or
-    // owned by the caller) and the signal (a caller abort after being served
-    // stale must not kill the shared refresh).
-    const bgOpts = {
-      ...opts,
-      headers: conditionalHeaders(key.headers ?? {}, entry),
-      body: null,
-      signal: undefined,
-    }
-    dispatch(
-      bgOpts,
-      new RevalidationHandler(key, entry, bgOpts, {
-        store,
-        logger: opts.logger,
-        handler: silentHandler,
-        allowStaleOnError: false,
-        noStore: false,
-        cacheOpts: cacheOptsOf(opts),
-      }),
-    )
-  } catch (err) {
-    done()
-    opts.logger?.debug({ err }, 'cache: background revalidation failed')
   }
 }
 
@@ -1430,11 +1070,7 @@ export default () => (dispatch) => (opts, handler) => {
   )
 }
 
-/**
- * @param {'hit' | 'miss'} [result] lookup outcome for the undici:cache doc
- * @param {string | null} [reason]
- */
-function serveFromCache(entry, opts, handler, write, url, lookupMs, result = 'hit', reason = null) {
+export function serveFromCache(entry, opts, handler) {
   const { statusCode, trailers, body } = entry
 
   let headers = entry.headers
@@ -1543,7 +1179,7 @@ function weakMatch(ifNoneMatch, etag) {
  * @param {string|any} etag
  * @returns {boolean}
  */
-function isEtagUsable(etag) {
+export function isEtagUsable(etag) {
   if (typeof etag !== 'string') {
     return false
   }
